requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.8.0
requests==2.32.5
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
bcrypt==4.3.0
SQLAlchemy==2.0.43
psutil==5.9.8
